Commit first draft of badge.rs
This module describes the Badge struct, which holds all necessary data for
constructing a badge. This file also includes many necessary functions that
aren't strictly underneath the Badge struct's perview, such as width derivation
of the left and right sides of the badge depending on the text used.

This file also utilizes Askama templating library as a way to inject badge data
into a .xml.j2 template, resulting in a proper svg with all the right data.
This template derivation and injection via Askama
allows for flexible implementation of badges down the road. This being said...
this templating probably belongs in its own file, along with the dimension
derivation.