Add the first SVG template for the "flat" badge.
This template is originally part of Google's pybadges
project (https://github.com/google/pybadges) and will be credited as such
eventually. I have modified it to simplify the logic (since Askama doesn't do OR
or AND statements like that) and clarify the named variables. I think I also
fixed a copy-paste bug that was in the pybadges repo... but I'll have to check
that later.