Change the *_color fields from Rgba to String.
This change was done in response to the error type BadgeError::ColorNotValid.
The validation of these colors was being done in two separate places for the
command line and the library; by making them String types, I can route both
paths through the same validation function, and propagate a single error type
from a single origin.