Use `include_bytes` to import font files for compatibility
It seems that one can't use file paths when importing assets into a Rust
program. The binary and the source files are distributed when installed through
Cargo; this makes asset paths invalid. In our case, the fonts just couldn't be
found.

`include_bytes` seems to get around this. However, to make sure that there is no
OS platform weirdness, the attets have to be in the same directory as the file
that calls them.